feat: enable plugin-manager for electron clients
this also adds support for setting the default command context via an env var

Fixes #2935
