+++ plugins/plugin-manager/src/test/manager/plugin.ts
